fix: load env files from stack directory
